fix(auth): harden SSO setup
