+++ apps/packages/product-domain/src/settings/cloud-settings.test.ts
@@ -32,7 +32,6 @@ describe("cloud settings section model", () => {
       "account",
       "environments",
       "organization",
-      "sso",
       "billing",
     ]);
   });
