+++ server/proliferate/integrations/sso/__init__.py
@@ -1,2 +1,1 @@
 """SSO protocol integrations."""
-
