+++ server/proliferate/server/organizations/sso/__init__.py
@@ -1,2 +1,1 @@
 """Organization SSO administration."""
-
